optimize: 剧集序号修改为浮点数

+++ server/src/main/java/run/ikaros/server/store/repository/EpisodeRepository.java
@@ -17,5 +17,5 @@ public interface EpisodeRepository extends R2dbcRepository<EpisodeEntity, Long> 
     Mono<Void> deleteAllBySubjectId(Long subjectId);
 
     Mono<EpisodeEntity> findBySubjectIdAndGroupAndSequence(Long subjectId, EpisodeGroup group,
-                                                           Integer sequence);
+                                                           Float sequence);
 }
